mcp_servers/python/servers/MYSQL_MCP_SERVER/mysql-mcp-server/pyproject.toml
[project]
name = "mysql-mcp-server"
version = "0.2.2"
description = "A Model Context Protocol (MCP) server that enables secure interaction with MySQL databases. This server allows AI assistants to list tables, read data, and execute SQL queries through a controlled interface, making database exploration and analysis safer and more structured."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "mcp>=1.0.0",
    "mysql-connector-python>=9.1.0",
]
[[project.authors]]
name = "Dana K. Williams"
email = "dana_w@designcomputer.com"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.scripts]
mysql-mcp-server = "mysql_mcp_server:main"
